requirements.txt
click==6.7
Flask==1.0.2
Flask-Cors==3.0.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pymongo==3.7.1
PyYAML==3.13
six==1.11.0
Werkzeug==0.14.1
